rebase: apply and cleanup autostash when rebase fails to start
If "git rebase" fails to start after stashing the user's uncommitted
changes then it forgets to restore the stashed changes and remove state
directory. To make matters worse running "git rebase --abort" to apply
the stashed changes and cleanup the state directory fails because the
state directory only contains the "autostash" file and is missing the
"head-name" and "onto" required by read_basic_state().

Fix this by applying the autostash and removing the state directory if
the pre-rebase hook or initial checkout fail. New tests are added to
check the autostash is applied and the state directory is removed if the
rebase fails to start. Checks are also added to some existing tests in
order to ensure there is no state directory left behind when a rebase
fails to start and no autostash has been created.

Reported-by: Brian Lyles <brianmlyles@gmail.com>
Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>