mergetool: set diff flag for vimdiff
The 'diff' flag is set when invoking vim with the '-d' option or as
'vimdiff'. However, the flag is not set before processing the .vimrc
file when vim is invoked via `git mergetool` since some point. This
can cause issues if the user has configurations in their vimrc that
depend on the 'diff' flag.

When using `git mergetool` for vimdiff, git generates commandline
parameters using the gen_cmd function in the mergetool/vimdiff script.
Commit 0041797449 removes the '-d' flag from the generated command.
Simply adding the flag back would ruin the layout, as the '-d' option
also tries to generate window layouts apart from the custom layout
commands of git.

The 'diff' flag is eventually set when 'diffthis' in git's layout
command is run. Unfortunately, these layout commands are specified with
'-c', which gets executed after the vimrc file processing. We can use
the '--cmd' option instead to make them execute before the vimrc, a
feature provided since vim-6-0ah. However, constructing the layout
before vimrc might change the behavior. So, I added '--cmd' only to turn
the 'diff' flag on.

Signed-off-by: Jeongho Lee <bivoje49@gmail.com>