Fix minor memory leak found by LeakSanitizer.
The callers of merge_recursive() and merge_ort_recursive() expects the
commit list passed in as the merge_bases parameter to be fully
consumed by the function and does not free it when the function
returns.  In normal cases, the commit list does get consumed, but when
the function returns early upon encountering an error, it forgets to
clean it up.

Fix this by freeing the list in the code paths for error returns.

Signed-off-by: Kevin Backhouse <kevinbackhouse@github.com>