rerere: allow find_conflict() to record found conflict type per path
The check_one_conflict() function signals the kind of conflict it
found to its caller with defined constants.  We would want to carry
it up to its caller, find_conflict() and store it in the list of
found conflicts.

Redefine these conflict_type constants as distinct "void *"
pointers, which is compatible with the type of .util member of
string_list_item, and store it in the .util member in the string
list the find_conflict() helper function returns.

Also switch from including only conflicts of the THREE_STAGED kind
to including anything but the PUNTED kind (i.e. those that "rerere"
will *never* handle), to prepare us to handle more conflict types
in the future.

Signed-off-by: Junio C Hamano <gitster@pobox.com>